chore: clean up copy texts
